# Conflicts:
#	hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/snapshot/OmSnapshotLocalDataManager.java

+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/snapshot/OmSnapshotLocalDataManager.java
@@ -423,6 +423,44 @@ public class OmSnapshotLocalDataManager implements AutoCloseable {
     }
   }
 
+  private synchronized void upsertNode(UUID snapshotId, SnapshotVersionsMeta snapshotVersions) throws IOException {
+    SnapshotVersionsMeta existingSnapVersions = getVersionNodeMap().remove(snapshotId);
+    Map<Integer, LocalDataVersionNode> existingVersions = existingSnapVersions == null ? Collections.emptyMap() :
+        existingSnapVersions.getSnapshotVersions();
+    Map<Integer, LocalDataVersionNode> newVersions = snapshotVersions.getSnapshotVersions();
+    Map<Integer, Set<LocalDataVersionNode>> predecessors = new HashMap<>();
+    boolean versionsRemoved = false;
+    // Track all predecessors of the existing versions and remove the node from the graph.
+    for (Map.Entry<Integer, LocalDataVersionNode> existingVersion : existingVersions.entrySet()) {
+      LocalDataVersionNode existingVersionNode = existingVersion.getValue();
+      predecessors.put(existingVersion.getKey(), localDataGraph.predecessors(existingVersionNode));
+      versionsRemoved = versionsRemoved || !newVersions.containsKey(existingVersion.getKey());
+      localDataGraph.removeNode(existingVersionNode);
+    }
+
+    // Add the nodes to be added in the graph and map.
+    addSnapshotVersionMeta(snapshotId, snapshotVersions);
+    // Reconnect all the predecessors for existing nodes.
+    for (Map.Entry<Integer, LocalDataVersionNode> entry : newVersions.entrySet()) {
+      for (LocalDataVersionNode predecessor : predecessors.getOrDefault(entry.getKey(), Collections.emptySet())) {
+        localDataGraph.putEdge(predecessor, entry.getValue());
+      }
+    }
+    if (existingSnapVersions != null) {
+      // The previous snapshotId could have become an orphan entry or could have orphan versions.(In case of
+      // version removals)
+      if (versionsRemoved || !Objects.equals(existingSnapVersions.getPreviousSnapshotId(),
+          snapshotVersions.getPreviousSnapshotId())) {
+        increamentOrphanCheckCount(existingSnapVersions.getPreviousSnapshotId());
+      }
+      // If the version is also updated it could mean that there could be some orphan version present within the
+      // same snapshot.
+      if (existingSnapVersions.getVersion() != snapshotVersions.getVersion()) {
+        increamentOrphanCheckCount(snapshotId);
+      }
+    }
+  }
+
   /**
    * The ReadableOmSnapshotLocalDataProvider class is responsible for managing the
    * access and initialization of local snapshot data in a thread-safe manner.
@@ -662,6 +700,8 @@ public class OmSnapshotLocalDataManager implements AutoCloseable {
    */
   public final class WritableOmSnapshotLocalDataProvider extends ReadableOmSnapshotLocalDataProvider {
 
+    private boolean dirty;
+
     private WritableOmSnapshotLocalDataProvider(UUID snapshotId) throws IOException {
       super(snapshotId, false);
       fullLock.readLock().lock();
@@ -681,6 +721,7 @@ public class OmSnapshotLocalDataManager implements AutoCloseable {
     private SnapshotVersionsMeta validateModification(OmSnapshotLocalData snapshotLocalData)
         throws IOException {
       SnapshotVersionsMeta versionsToBeAdded = new SnapshotVersionsMeta(snapshotLocalData);
+      SnapshotVersionsMeta existingVersionsMeta = getVersionNodeMap().get(snapshotLocalData.getSnapshotId());
       for (LocalDataVersionNode node : versionsToBeAdded.getSnapshotVersions().values()) {
         validateVersionAddition(node);
       }
@@ -692,50 +733,15 @@ public class OmSnapshotLocalDataManager implements AutoCloseable {
           validateVersionRemoval(snapshotId, entry.getKey());
         }
       }
-      SnapshotVersionsMeta existingVersionMeta = getVersionNodeMap().get(snapshotId);
-      // Set the needsDefrag if the new previous snapshotId is different from the existing one or if this is a new
-      // snapshot yaml file.
-      snapshotLocalData.setNeedsDefrag(existingVersionMeta == null
-          || !Objects.equals(existingVersionMeta.getPreviousSnapshotId(), snapshotLocalData.getPreviousSnapshotId()));
+      // Set Dirty if the snapshot doesn't exist or previousSnapshotId has changed.
+      if (existingVersionsMeta == null || !Objects.equals(versionsToBeAdded.getPreviousSnapshotId(),
+          existingVersionsMeta.getPreviousSnapshotId())) {
+        setDirty();
+        // Set the needsDefrag if the new previous snapshotId is different from the existing one or if this is a new
+        // snapshot yaml file.
+        snapshotLocalData.setNeedsDefrag(true);
+      }
       return versionsToBeAdded;
-    }
-
-    private synchronized void upsertNode(UUID snapshotId, SnapshotVersionsMeta snapshotVersions) throws IOException {
-      SnapshotVersionsMeta existingSnapVersions = getVersionNodeMap().remove(snapshotId);
-      Map<Integer, LocalDataVersionNode> existingVersions = existingSnapVersions == null ? Collections.emptyMap() :
-          existingSnapVersions.getSnapshotVersions();
-      Map<Integer, LocalDataVersionNode> newVersions = snapshotVersions.getSnapshotVersions();
-      Map<Integer, Set<LocalDataVersionNode>> predecessors = new HashMap<>();
-      boolean versionsRemoved = false;
-      // Track all predecessors of the existing versions and remove the node from the graph.
-      for (Map.Entry<Integer, LocalDataVersionNode> existingVersion : existingVersions.entrySet()) {
-        LocalDataVersionNode existingVersionNode = existingVersion.getValue();
-        predecessors.put(existingVersion.getKey(), localDataGraph.predecessors(existingVersionNode));
-        versionsRemoved = versionsRemoved || !newVersions.containsKey(existingVersion.getKey());
-        localDataGraph.removeNode(existingVersionNode);
-      }
-
-      // Add the nodes to be added in the graph and map.
-      addSnapshotVersionMeta(snapshotId, snapshotVersions);
-      // Reconnect all the predecessors for existing nodes.
-      for (Map.Entry<Integer, LocalDataVersionNode> entry : newVersions.entrySet()) {
-        for (LocalDataVersionNode predecessor : predecessors.getOrDefault(entry.getKey(), Collections.emptySet())) {
-          localDataGraph.putEdge(predecessor, entry.getValue());
-        }
-      }
-      if (existingSnapVersions != null) {
-        // The previous snapshotId could have become an orphan entry or could have orphan versions.(In case of
-        // version removals)
-        if (versionsRemoved || !Objects.equals(existingSnapVersions.getPreviousSnapshotId(),
-            snapshotVersions.getPreviousSnapshotId())) {
-          increamentOrphanCheckCount(existingSnapVersions.getPreviousSnapshotId());
-        }
-        // If the version is also updated it could mean that there could be some orphan version present within the
-        // same snapshot.
-        if (existingSnapVersions.getVersion() != snapshotVersions.getVersion()) {
-          increamentOrphanCheckCount(snapshotId);
-        }
-      }
     }
 
     public void addSnapshotVersion(RDBStore snapshotStore) throws IOException {
@@ -743,33 +749,59 @@ public class OmSnapshotLocalDataManager implements AutoCloseable {
       OmSnapshotLocalData previousSnapshotLocalData = getPreviousSnapshotLocalData();
       this.getSnapshotLocalData().addVersionSSTFileInfos(sstFiles, previousSnapshotLocalData == null ? 0 :
           previousSnapshotLocalData.getVersion());
+      // Set Dirty if a version is added.
+      setDirty();
+    }
+
+    public void removeVersion(int version) {
+      this.getSnapshotLocalData().removeVersionSSTFileInfos(version);
+      // Set Dirty if a version is removed.
+      setDirty();
     }
 
     public synchronized void commit() throws IOException {
+      // Validate modification and commit the changes.
       SnapshotVersionsMeta localDataVersionNodes = validateModification(super.snapshotLocalData);
-      String filePath = getSnapshotLocalPropertyYamlPath(super.snapshotId);
-      File snapshotLocalDataFile = new File(filePath);
-      if (!localDataVersionNodes.getSnapshotVersions().isEmpty()) {
-        String tmpFilePath = filePath + ".tmp";
-        File tmpFile = new File(tmpFilePath);
-        boolean tmpFileExists = tmpFile.exists();
-        if (tmpFileExists) {
-          tmpFileExists = !tmpFile.delete();
+      // Need to update the disk state if and only if the dirty bit is set.
+      if (isDirty()) {
+        String filePath = getSnapshotLocalPropertyYamlPath(super.snapshotId);
+        File snapshotLocalDataFile = new File(filePath);
+        if (!localDataVersionNodes.getSnapshotVersions().isEmpty()) {
+          String tmpFilePath = filePath + ".tmp";
+          File tmpFile = new File(tmpFilePath);
+          boolean tmpFileExists = tmpFile.exists();
+          if (tmpFileExists) {
+            tmpFileExists = !tmpFile.delete();
+          }
+          if (tmpFileExists) {
+            throw new IOException("Unable to delete tmp file " + tmpFilePath);
+          }
+          snapshotLocalDataSerializer.save(new File(tmpFilePath), super.snapshotLocalData);
+          Files.move(tmpFile.toPath(), Paths.get(filePath), StandardCopyOption.ATOMIC_MOVE,
+              StandardCopyOption.REPLACE_EXISTING);
+        } else if (snapshotLocalDataFile.exists()) {
+          LOG.info("Deleting Yaml file corresponding to snapshotId: {} in path : {}",
+              super.snapshotId, snapshotLocalDataFile.getAbsolutePath());
+          if (snapshotLocalDataFile.delete()) {
+            throw new IOException("Unable to delete file " + snapshotLocalDataFile.getAbsolutePath());
+          }
         }
-        if (tmpFileExists) {
-          throw new IOException("Unable to delete tmp file " + tmpFilePath);
-        }
-        snapshotLocalDataSerializer.save(new File(tmpFilePath), super.snapshotLocalData);
-        Files.move(tmpFile.toPath(), Paths.get(filePath), StandardCopyOption.ATOMIC_MOVE,
-            StandardCopyOption.REPLACE_EXISTING);
-      } else if (snapshotLocalDataFile.exists()) {
-        LOG.info("Deleting Yaml file corresponding to snapshotId: {} in path : {}",
-            super.snapshotId, snapshotLocalDataFile.getAbsolutePath());
-        if (snapshotLocalDataFile.delete()) {
-          throw new IOException("Unable to delete file " + snapshotLocalDataFile.getAbsolutePath());
-        }
+        upsertNode(super.snapshotId, localDataVersionNodes);
+        // Reset dirty bit
+        resetDirty();
       }
-      upsertNode(super.snapshotId, localDataVersionNodes);
+    }
+
+    private void setDirty() {
+      dirty = true;
+    }
+
+    private void resetDirty() {
+      dirty = false;
+    }
+
+    private boolean isDirty() {
+      return dirty;
     }
 
     @Override
